Added judicial api client and refactored
