+++ src/functionalTest/java/uk/gov/hmcts/reform/judicialapi/RetrieveJudicialRolesTest.java
@@ -12,6 +12,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.http.HttpStatus;
 import org.springframework.test.context.ActiveProfiles;
+import uk.gov.hmcts.reform.judicialapi.AuthorizationFunctionalTest;
 
 import java.util.Map;
 
@@ -25,7 +26,7 @@ public class RetrieveJudicialRolesTest extends AuthorizationFunctionalTest {
         //Given I am a user with the appropriate rights to retrieve list of all Judicial Roles
         //When I search for the list of all Judicial Roles
         //Then I should be able to see the list of Judicial roles
-        Map<String, Object> response = retrieveAllJudicialRoles("caseworker", HttpStatus.OK);
+        Map<String, Object> response = judicialApiClient.retrieveAllJudicialRoles("caseworker", HttpStatus.OK);
         assertThat(response).isNotNull();
         assertThat(response).containsValues("Magistrate", "Advisory Committee Member - Magistrate");
     }
@@ -35,30 +36,9 @@ public class RetrieveJudicialRolesTest extends AuthorizationFunctionalTest {
         //Given I am a user without the appropriate rights to retrieve list of all Judicial Roles
         //When I search for the list of all Judicial Roles
         //Then I should get a failure outcome as I do not have the appropriate rights for accessing role information
-        Map<String, Object> response = retrieveAllJudicialRoles("pui-user-manager", HttpStatus.FORBIDDEN);
+        Map<String, Object> response = judicialApiClient.retrieveAllJudicialRoles("pui-user-manager", HttpStatus.FORBIDDEN);
         assertThat(response.get("errorMessage")).isNotNull();
         assertThat(response.get("errorMessage")).isEqualTo("9 : Access Denied");
     }
 
-    private RequestSpecification withUnauthenticatedRequest() {
-        return SerenityRest.given()
-                .relaxedHTTPSValidation()
-                .baseUri(judicialApiUrl)
-                .header("Content-Type", APPLICATION_JSON_UTF8_VALUE)
-                .header("Accepts", APPLICATION_JSON_UTF8_VALUE);
-    }
-
-    public Map<String, Object> retrieveAllJudicialRoles(String roleOfAccessor, HttpStatus expectedStatus) {
-        Response response = withUnauthenticatedRequest()
-                .body("")
-                .get("/refdata/v1/judicial/roles")
-                .andReturn();
-
-
-        response.then()
-                .assertThat()
-                .statusCode(expectedStatus.value());
-
-        return response.body().as(Map.class);
-    }
 }
