+++ src/functionalTest/java/uk/gov/hmcts/reform/judicialapi/client/JudicialApiClient.java
@@ -1,0 +1,68 @@
+package uk.gov.hmcts.reform.judicialapi.client;
+
+import io.restassured.response.Response;
+import io.restassured.specification.RequestSpecification;
+import lombok.extern.slf4j.Slf4j;
+import net.serenitybdd.rest.SerenityRest;
+import org.springframework.http.HttpStatus;
+
+import java.util.Map;
+
+import static org.springframework.http.HttpStatus.OK;
+import static org.springframework.http.MediaType.APPLICATION_JSON_UTF8_VALUE;
+
+@Slf4j
+public class JudicialApiClient {
+
+    private final String judicialApiUrl;
+    private final String s2sToken;
+
+    public JudicialApiClient(String judicialApiUrl, String s2sToken) {
+        this.judicialApiUrl = judicialApiUrl;
+        this.s2sToken = s2sToken;
+    }
+
+    public String getWelcomePage() {
+        return withUnauthenticatedRequest()
+                .get("/")
+                .then()
+                .statusCode(OK.value())
+                .and()
+                .extract()
+                .body()
+                .asString();
+    }
+
+    public String getHealthPage() {
+        return withUnauthenticatedRequest()
+                .get("/health")
+                .then()
+                .statusCode(OK.value())
+                .and()
+                .extract()
+                .body()
+                .asString();
+    }
+
+    private RequestSpecification withUnauthenticatedRequest() {
+        return SerenityRest.given()
+                .relaxedHTTPSValidation()
+                .baseUri(judicialApiUrl)
+                .header("Content-Type", APPLICATION_JSON_UTF8_VALUE)
+                .header("Accepts", APPLICATION_JSON_UTF8_VALUE);
+    }
+
+    public Map<String, Object> retrieveAllJudicialRoles(String roleOfAccessor, HttpStatus expectedStatus) {
+        Response response = withUnauthenticatedRequest()
+                .body("")
+                .get("/refdata/v1/judicial/roles")
+                .andReturn();
+
+
+        response.then()
+                .assertThat()
+                .statusCode(expectedStatus.value());
+
+        return response.body().as(Map.class);
+    }
+}
